+++ content/authorization/authorizing-api-calls/index.mdx
@@ -69,7 +69,7 @@ Personal Access Tokens use the <a href="https://developer.mozilla.org/en-US/docs
 
 Use:
 
-- **Account ID** as the **username** for the [Agent Chat API v3](/messaging/agent-chat-api/), [Configuration API v3](/management/configuration-api/), [LiveChat Accounts API](/authorization/livechat-accounts-api/), and [HelpDesk API](https://api.helpdesk.com/docs)
+- **Account ID** as the **username** for the [Agent Chat API v3](/messaging/agent-chat-api/), [Configuration API v3](/management/configuration-api/), [Global Accounts API](/authorization/global-accounts-api/), and [HelpDesk API](https://api.helpdesk.com/docs)
 - **Entity ID (your LiveChat login)** as the **username** for the [Configuration API v2](/management/configuration-api/v2.0/) (the legacy version)
 - **PAT** as the **password** for the APIs v2 and v3
