+++ content/authorization/global-accounts-api/index.mdx
@@ -1,10 +1,10 @@
 ---
 weight: 20
 category: "authorization"
-subcategory: "livechat-accounts-api"
-title: LiveChat Account API
+subcategory: "global-accounts-api"
+title: Global Accounts API
 tagline: "Find out how you to configure chat webhooks."
 desc: "Use chat webhooks to get notified about the events happening during chats, for example, when there is an incoming chat, when someone sends a message, or when an agent tags a chat."
 ---
 
-<Redoc name="livechat-accounts-api" />
+<Redoc name="global-accounts-api" />
