+++ content/authorization/global-accounts/index.mdx
@@ -1,7 +1,7 @@
 ---
 weight: 10
 category: 'authorization'
-subcategory: "livechat-accounts"
+subcategory: "global-accounts"
 title: 'Global Accounts'
 tagline: 'Learn about LiveChat Global Accounts.'
 desc: 'Learn about LiveChat Global Accounts, a service where you can configure options and resources common for all products.'
@@ -68,4 +68,4 @@ With the introduction of Global Accounts, we removed two roles from the LiveChat
 
 # Global Accounts API
 
-You can find the Global Accounts API specification [in this document](/authorization/livechat-accounts-api/).
+You can find the Global Accounts API specification [in this document](/authorization/global-accounts-api/).
