+++ content/authorization/index.mdx
@@ -13,8 +13,8 @@ To interact with the LiveChat APIs, you need to provide an access token. The app
 
 Before you jump into implementing an authorization flow in your app, make sure you understand what Global Accounts is.
 
-<SectionLink to={"/authorization/livechat-accounts/"}>Global Accounts</SectionLink>
-<SectionLink to={"/authorization/livechat-accounts-api/"}>Global Accounts API</SectionLink>
+<SectionLink to={"/authorization/global-accounts/"}>Global Accounts</SectionLink>
+<SectionLink to={"/authorization/global-accounts-api/"}>Global Accounts API</SectionLink>
 
 ## Authorization tutorials
