+++ content/categories.json
@@ -33,15 +33,15 @@
       "color": "67, 132, 245",
       "items": [
         {
-          "slug": "livechat-accounts",
-          "title": "LiveChat Accounts",
-          "desc": "Learn about LiveChat Accounts, a service where you can configure options and resources common for all products.",
+          "slug": "global-accounts",
+          "title": "Global Accounts",
+          "desc": "Learn about Global Accounts, a service where you can configure options and resources common for all products.",
           "color": "67, 132, 245"
         },
         {
-          "slug": "livechat-accounts-api",
-          "title": "LiveChat Accounts API",
-          "desc": "LiveChat Accounts API reference.",
+          "slug": "global-accounts-api",
+          "title": "Global Accounts API",
+          "desc": "Global Accounts API reference.",
           "color": "67, 132, 245"
         },
         {
