+++ content/extending-agent-app/product-cards/index.mdx
@@ -42,7 +42,7 @@ To make Product Cards work with your store, you need to prepare own backend serv
 
 ## Frontend
 
-The frontend flow (presented in the diagram above) is based on communication between the app, your backend service, and [LiveChat Accounts](/authorization/livechat-accounts):
+The frontend flow (presented in the diagram above) is based on communication between the app, your backend service, and [Global Accounts](/authorization/global-accounts):
 
 1. The app uses [Sign in with LiveChat](/authorization/sign-in-with-livechat/) to acquire the access token. This token will be used to [authorize requests](/authorization/authorizing-api-calls/#validating-the-access-token) in your backend service for requests below (via `Authorization` header).
 2. Then, the app fetches the `/stores` endpoint to get the list of available stores.
