+++ src/configs/redoc/global-accounts-api/config.js
