+++ src/configs/redoc/global-accounts-api/spec.yml
